moved graph crawl helper methods to MetadataService
